LayoutTests/fast/dom/window-open-ephemeral-expected.txt

Tests that window.open() is working in ephemeral sessions

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !!w is true
PASS successfullyParsed is true

TEST COMPLETE
